Module 1 #1

+++ minitorch/operators.py
@@ -25,7 +25,7 @@ def add(x: float, y: float) -> float:
 
 def neg(x: float) -> float:
     "$f(x) = -x$"
-    return -x
+    return (float)(-x)
 
 def lt(x: float, y: float) -> float:
     "$f(x) =$ 1.0 if x is less than y else 0.0"
@@ -83,7 +83,7 @@ def relu(x: float) -> float:
     if x > 0:
         return x
     else:
-        return 0
+        return 0.0
 
 EPS = 1e-6
 
@@ -118,7 +118,7 @@ def relu_back(x: float, d: float) -> float:
     if x > 0:
         return d
     else:
-        return 0
+        return 0.0
 
 
 # ## Task 0.3
